+++ packages/atomic/src/components/search/atomic-did-you-mean/atomic-did-you-mean.stories.tsx
@@ -16,5 +16,5 @@ const {defaultModuleExport, exportedStory} = defaultStory(
   }
 );
 
-export default defaultModuleExport;
+export default {...defaultModuleExport, title: 'Atomic/DidYouMean'};
 export const DefaultDidYouMean = exportedStory;