+++ packages/atomic/src/components/search/result-template-components/atomic-result-number/atomic-result-number.stories.tsx
@@ -21,5 +21,8 @@ const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
   }
 );
 
-export default defaultModuleExport;
+export default {
+  ...defaultModuleExport,
+  title: 'Atomic/ResultList/ResultNumber',
+};
 export const DefaultResultNumber = exportedStory;