+++ packages/atomic/src/components/search/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results.stories.tsx
@@ -70,5 +70,8 @@ const {defaultModuleExport, exportedStory} = defaultStory(
   }
 );
 
-export default defaultModuleExport;
 export const DefaultInstantResults = exportedStory;
+export default {
+  ...defaultModuleExport,
+  title: 'Atomic/Searchbox/InstantResults',
+};