+++ packages/headless/src/features/did-you-mean/did-you-mean-insight-analytics-actions.test.ts
@@ -1,4 +1,4 @@
-import * as CoveoAnalytics from 'coveo.analytics';
+//import {CoveoInsightClient} from 'coveo.analytics';
 import {buildMockInsightEngine} from '../../test/mock-engine';
 import {buildMockInsightState} from '../../test/mock-insight-state';
 import {
@@ -9,14 +9,17 @@ import {
 const mockLogDidYouMeanClick = jest.fn();
 const mockLogDidYouMeanAutomatic = jest.fn();
 
-const mockCoveoInsightClient = jest.fn(() => ({
-  disable: () => {},
-  logDidYouMeanClick: mockLogDidYouMeanClick,
-  logDidYouMeanAutomatic: mockLogDidYouMeanAutomatic,
-}));
+jest.mock('coveo.analytics', () => {
+  const mockCoveoInsightClient = jest.fn(() => ({
+    disable: () => {},
+    logDidYouMeanClick: mockLogDidYouMeanClick,
+    logDidYouMeanAutomatic: mockLogDidYouMeanAutomatic,
+  }));
 
-Object.defineProperty(CoveoAnalytics, 'CoveoInsightClient', {
-  value: mockCoveoInsightClient,
+  return {
+    CoveoInsightClient: mockCoveoInsightClient,
+    history: {HistoryStore: jest.fn()},
+  };
 });
 
 const exampleSubject = 'example subject';