+++ packages/headless/src/insight.index.ts
@@ -1,3 +1,6 @@
+import {polyfillCryptoNode} from './api/analytics/analytics-crypto-polyfill';
+
+polyfillCryptoNode();
 // 3rd Party Libraries
 export type {Unsubscribe, Middleware} from '@reduxjs/toolkit';
 
@@ -32,6 +35,7 @@ export * from './features/case-context/case-context-actions-loader';
 export * from './features/insight-search/insight-search-analytics-actions-loader';
 export * from './features/fields/fields-actions-loader';
 export * from './features/attached-results/attached-results-actions-loader';
+export * from './features/analytics/generic-analytics-actions-loader';
 
 // Controllers
 export type {
@@ -261,6 +265,16 @@ export type {InsightInterfaceState} from './features/insight-interface/insight-i
 
 export type {InsightInterface} from './controllers/insight-interface/insight-interface';
 export {buildInsightInterface} from './controllers/insight-interface/insight-interface';
+
+export type {
+  SmartSnippet,
+  SmartSnippetOptions,
+  SmartSnippetProps,
+  SmartSnippetState,
+  QuestionAnswerDocumentIdentifier,
+  SmartSnippetCore,
+} from './controllers/insight/smart-snippet/headless-insight-smart-snippet';
+export {buildSmartSnippet} from './controllers/insight/smart-snippet/headless-insight-smart-snippet';
 
 // Features
 export type {